streamlined erro handling and added timeout option
Error handling has been streamlined so that callbacks are invoked in
error cases in every method.

A request timeout  option has been added for backend requests. A timeout
passed to hte constructor is used as the global default while timeouts
set in method options are only used in the method call-